fix(control-plane): restore release gate health
